requirements.txt
django
djangorestframework
psycopg
celery
redis